Function peepmatic::compile_file

pub fn compile_file<TOperator>(
    filename: &Path
) -> Result<PeepholeOptimizations<TOperator>> where
    TOperator: Copy + Debug + Eq + Hash + for<'a> Parse<'a> + TypingRules + Into<NonZeroU32> + TryFrom<NonZeroU32>, 

Compile the given DSL file into a compact peephole optimizations automaton!

Example

use std::path::Path;

let peep_opts = peepmatic::compile_file::<cranelift_codegen::ir::Opcode>(
    Path::new("path/to/optimizations.peepmatic")
)?;

// Use the peephole optimizations or serialize them into bytes here...

Visualizing the Peephole Optimizer's Automaton

To visualize (or debug) the peephole optimizer's automaton, set the PEEPMATIC_DOT environment variable to a file path. A GraphViz Dot file showing the peephole optimizer's automaton will be written to that file path.