Function peepmatic::compile_str

pub fn compile_str<TOperator>(
    source: &str, 
    filename: &Path
) -> Result<PeepholeOptimizations<TOperator>> where
    TOperator: Copy + Debug + Eq + Hash + for<'a> Parse<'a> + TypingRules + Into<NonZeroU32> + TryFrom<NonZeroU32>, 

Compile the given DSL source text down into a compact peephole optimizations automaton.

This is like compile_file but you bring your own file I/O.

The filename parameter is used to provide better error messages.

Example

use std::path::Path;

let peep_opts = peepmatic::compile_str::<cranelift_codegen::ir::Opcode>(
    "
    (=> (iadd $x 0) $x)
    (=> (imul $x 0) 0)
    (=> (imul $x 1) $x)
    ",
    Path::new("my-optimizations"),
)?;

// Use the peephole optimizations or serialize them into bytes here...

Visualizing the Peephole Optimizer’s Automaton

To visualize (or debug) the peephole optimizer’s automaton, set the PEEPMATIC_DOT environment variable to a file path. A GraphViz Dot file showing the peephole optimizer’s automaton will be written to that file path.