Function peepmatic::match_in_same_order

pub fn match_in_same_order<TOperator>(opts: &mut Optimizations<TOperator>) where
    TOperator: Copy + Debug + Eq + Hash, 

Ensure that we emit match operations in a consistent order.

There are many linear optimizations, each of which have their own sequence of match operations that need to be tested. But when interpreting the automata against some instructions, we only perform a single sequence of match operations, and at any given moment, we only want one match operation to interpret next. This means that two optimizations that are next to each other in the sorting must have their shared prefixes diverge on an expected result edge, not on which match operation to preform next. And if they have zero shared prefix, then we need to create one, that immediately divereges on the expected result.

For example, consider these two patterns that don’t have any shared prefix:

(=> (iadd $x $y) ...)
(=> $C ...)

These produce the following linear match operations and expected results:

opcode @ 0 --iadd-->
is-const? @ 0 --true-->

In order to ensure that we only have one match operation to interpret at any given time when evaluating the automata, this pass transforms the second optimization so that it shares a prefix match operation, but diverges on the expected result:

opcode @ 0 --iadd-->
opcode @ 0 --(else)--> is-const? @ 0 --true-->