Function peepmatic::remove_unnecessary_nops

pub fn remove_unnecessary_nops<TOperator>(opts: &mut Optimizations<TOperator>) where
    TOperator: Copy + Debug + Eq + Hash, 

99.99% of nops are unnecessary; remove them.

They're only needed for when a LHS pattern is just a variable, and that's it. However, it is easier to have basically unused nop matching operations for the DSL's edge-cases than it is to try and statically eliminate their existence completely. So we just emit nop match operations for all variable patterns, and then in this post-processing pass, we fuse them and their actions with their preceding increment.