Trait peepmatic::Ast

pub trait Ast<'a, TOperator>: 'a + ChildNodes<'a, 'a, TOperator> + Span where
    DynAstRef<'a, TOperator>: From<&'a Self>,
    TOperator: 'a, { }

A trait implemented by all AST nodes.

All AST nodes can:

Enumerate their children via ChildNodes.

Give you the wast::Span where they were defined.

Be converted into a DynAstRef.

This trait is blanked implemented for everything that does those three things, and in practice those three thrings are all implemented by the derive(Ast) macro.